ALOOP Re-create donations.yml (#186)
* AbleGamers donation by ALoop

* Add files via upload

* Changed L to upper case to resolve the path

* Update donations.yml

* Add files via upload

* ALOOP August 2021 donation update

* Update donations.yml

* Update donations.yml

* Update donations.yml

* ALOOP Donation

Capitalize folder name and added a second donation

* Delete ALoop directory

* ALOOP donations

* Update donations.yml

* Delete 2021-08-14-ABG.png

* ALOOP Replaced donation

ALOOP Replaced donation with proper tax receipt

* ALOOP August donation confirmation

Yes, this is a recurring payment on my credit card. Receipt comes only one time, after that I received only a confirmation. Another MDP admin asked to update the receipt instead of confirmation. I did it on my second PR. Please check on the right side of the amount, it shows as 'Donation recurring monthly' , screenshot attached.

* Updated to re-run validation

* Update donation AbleGamers September

* Update donations.yml

* ALOOP September Donation

* Update donations.yml

* ALOOP September Donation

* ALOOP Update Donations

* Remove conflict files

* Remove conflict files

* Restore ALOOP payment for August

* ALOOP donation update

* Delete Donations.yml

* Re-create donations.yml

* ALOOP changed donations.yml

I still have an error on this, but don't see why and where is the conflict.

* Removed extra files and update donations.yml

Co-authored-by: ResponsibleStaking <79546146+ResponsibleStaking@users.noreply.github.com>
Co-authored-by: Kiriakos Krastillis <kiriakos@kind.software>