Add donation for epochs 325-326-327-328-329 (#317)
* Create donations.yml

* Add files via upload

* Add files via upload

* Update donations.yml

* Delete Certificazione Fiscale_20211010_310088.png

* Add files via upload

* Update donations.yml

* Update donations.yml

* Update donations.yml

* Update donations.yml

* Update donations.yml

* Update donations.yml

* Update donation.yml

We've made some mess with ADA/EUR/DOLLAR conversion since the payment has been made in €. Updated the dollar field with right conversion and added a custom field of the € payment amount.

* Update donations.yml

* Add new donation

ForestaMi invoices are now generated 50 days after the donation submission. To be in time with our donation schedule, we will push the receipt (available as soon as the donation is submitted), replacing it with the invoice as soon as generated. Thanks.

* Add donation #3 receipt 

This will be replaced with the invoice as soon as available (50 day after the donation submission). Thanks

* Add receipt #4

This file will be replaced with the invoice as soon as available (about 30 days)

* Add donation #4

* Update donations.yml

* Update donations.yml

* Update donations.yml

* Update donations.yml

* Update donations.yml

* Add files via upload

* Delete Certificazione_Fiscale_20220210.png

* Delete Certificazione_Fiscale_20220313.png

* Uploaded the invoices

This commit replaces the last two temporary receipts with the final ones. These take a long time to deliver from the project we are supporting. Thanks :)

* Update donations.yml

* Delete Certificazione Fiscale_20220313_347880.png

* Delete Certificazione Fiscale_20220210_340022.png

* Add files via upload

* Update file name references

* Add temporary receipt 

This will be replaced with the invoice ad soon as released by the fund.

* Add receipt

This will be replaced with the invoice as soon as released by the fund.

* Add donation for epochs  325-326-327-328-329

Co-authored-by: Garrett Seymour <nova4005@users.noreply.github.com>
Co-authored-by: ALOOP Stake Pool <85917973+aloopmarin007@users.noreply.github.com>